changes for continualtest; test changes only

+++ tests/continualtest.sh
@@ -12,16 +12,19 @@ This script runs the command line tests
 OPTIONS:
    -h      Show this message
    -i      Number of iterations
+   -d      Directory name (default is fusedav)
+   -p      pid of fusedav instance
    -v      Verbose
 EOF
 }
 
 # do 64 rounds by default
+pid=0
 iters=64
 verbose=0
 # assuming we are using base fusedav; but allow for override
 fusedavdir="fusedav"
-while getopts "hi:d:v" OPTION
+while getopts "hi:d:p:v" OPTION
 do
      case $OPTION in
          h)
@@ -33,6 +36,9 @@ do
              ;;
          d)
              fusedavdir=$OPTARG
+             ;;
+         p)
+             pid=$OPTARG
              ;;
          v)
              verbose=1
@@ -52,45 +58,51 @@ if [ $bname != 'files' ]; then
     echo "Expected base directory to be called files"
 fi
 bindingdir=$(dirname $curdir)
+
+if [ $pid -eq 0 ]; then
+    echo "Requires pid"
+    exit 1
+fi
     
 pprof_out=$bindingdir/pprof.out
 rm -f $pprof_out
 iter=0
 while [ $iter -le $iters ]
 do
-    # get the pid
-    pid=$(ps aux | grep mount.$fusedavdir | grep -v grep | sort -nrk 6 | awk '{printf "%5d\n", $2}')
-    
     # get the current memory use
-    res=$(ps aux | grep mount.$fusedavdir | grep -v grep | sort -nrk 6 | awk '{printf "%5d %d\n", $2, $6}')
+    res=$(ps aux | grep mount.$fusedavdir | grep -v grep | grep $pid | awk '{printf "%5d %d\n", $2, $6}')
     echo "$iter: before make: $res"
     echo "$iter: before make: $res" >> $pprof_out
     
-    echo "$iter: make -f /opt/$fusedavdir/tests/Makefile testdir=/opt/$fusedavdir/tests"
-    echo "$iter: make -f /opt/$fusedavdir/tests/Makefile testdir=/opt/$fusedavdir/tests" >> $pprof_out
+    echo "$iter: make -f /opt/$fusedavdir/tests/Makefile run-continual-tests testdir=/opt/$fusedavdir/tests"
+    echo "$iter: make -f /opt/$fusedavdir/tests/Makefile run-continual-tests testdir=/opt/$fusedavdir/tests" >> $pprof_out
     
     # run the tests
-    make -f /opt/$fusedavdir/tests/Makefile testdir=/opt/$fusedavdir/tests
+    make -f /opt/$fusedavdir/tests/Makefile run-continual-tests testdir=/opt/$fusedavdir/tests
 
-    # get the most recent heap to use as base to pprof
-    prevheap=$(ls -Art /var/tmp/*.heap | tail -n 1)
+    if [ $iter -gt 0 ]; then
+        # get the most recent heap to use as base to pprof
+        prevheap=$(ls -Art /var/tmp/*${pid}*.heap | tail -n 1)
+    fi
 
     # Send our mallctl signal to fusedav; this will dump a jemalloc heap file
     echo "kill -SIGUSR2 $pid"
     kill -SIGUSR2 $pid
     
     # get the heap file just created by the SIGUSR2
-    newheap=$(ls -Art /var/tmp/*.heap | tail -n 1)
+    newheap=$(ls -Art /var/tmp/*${pid}*.heap | tail -n 1)
     
-    # do a pprof comparison between latest and previous heap filees
-    echo "pprof --text --lines --inuse_space --base=$prevheap /opt/$fusedavdir/src/fusedav $newheap >> $pprof_out 2>&1"
-    pprof --text --lines --inuse_space --base=$prevheap /opt/$fusedavdir/src/fusedav $newheap >> $pprof_out 2>&1
+    if [ $iter -gt 0 ]; then
+        # do a pprof comparison between latest and previous heap filees
+        echo "pprof --text --lines --inuse_space --base=$prevheap /opt/$fusedavdir/src/fusedav $newheap >> $pprof_out 2>&1"
+        pprof --text --lines --inuse_space --base=$prevheap /opt/$fusedavdir/src/fusedav $newheap >> $pprof_out 2>&1
+    fi
     
-    res=$(ps aux | grep mount.$fusedavdir | grep -v grep | sort -nrk 6 | awk '{printf "%5d %d\n", $2, $6}')
+    res=$(ps aux | grep mount.$fusedavdir | grep -v grep | grep $pid | awk '{printf "%5d %d\n", $2, $6}')
     echo "$iter: after make: $res"
     
     # leave some time for things to settle out before next round
-    sleep 60
+    # no, dont sleep 60
     iter=$((iter + 1))
 done
